requirements-dev.txt
sphinx
sphinx-click

flake8
coverage
pytest
pytest-cov
pytest-flake8
ptpython
cprofilev
